requirements.txt
# 데이터 크롤링 및 정렬
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
h11==0.14.0
idna==3.10
numpy==2.2.1
outcome==1.3.0.post0
pandas==2.2.3
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0

# AI 학습 및 최적화
unsloth
transformers
torch
datasets
trl
tqdm

# AI 추론 및 Gradio 기반 프론트엔드
torchvision
torchaudio
pymongo
scikit-learn
fastapi
uvicorn
pyngrok
nest-asyncio
pillow
python-multipart
pydantic

# 안드로이드 개발 관련 (PyPI에서 직접 설치 불가, 별도 설정 필요)
# agp==8.5.0
# annotation==1.9.1
# converterGson==2.6.4
# junit==4.13.2
# junitVersion==1.2.0
# espressoCore==3.6.0
# appcompat==1.7.0
# material==1.12.0
# activity==1.9.0
# constraintlayout==2.1.4
# retrofit==2.6.4
# visionCommon==17.3.0
# playServicesTasks==18.2.0
# segmentationSelfie==16.0.0-beta6
# playServicesMlkitSubjectSegmentation==16.0.0-beta1
